# MQE: Fix native histogram bugs