@jhesketh 2024-08-30
This has a very slight negative impact to the benchmark due to copying the span values. However it appears acceptable to me:

<details>

```

                                                                           │  main.txt   │             compact.txt             │
                                                                           │   sec/op    │    sec/op     vs base               │
Query/nh_1,_instant_query/Mimir-8                                            791.7µ ± 7%    806.4µ ± 1%       ~ (p=0.093 n=6)
Query/nh_1_+_nh_1,_instant_query/Mimir-8                                     1.503m ± 1%    1.521m ± 1%  +1.17% (p=0.024 n=12)
Query/nh_1,_instant_query/Prometheus-8                                       809.6µ ± 1%    815.5µ ± 1%       ~ (p=0.132 n=6)
Query/nh_1_+_nh_1,_instant_query/Prometheus-8                                1.514m ± 0%    1.534m ± 2%  +1.34% (p=0.000 n=12)
Query/nh_1,_range_query_with_100_steps/Mimir-8                               919.3µ ± 4%    911.7µ ± 1%       ~ (p=0.589 n=6)
Query/nh_1_+_nh_1,_range_query_with_100_steps/Mimir-8                        1.726m ± 2%    1.725m ± 1%       ~ (p=0.887 n=12)
Query/nh_1,_range_query_with_100_steps/Prometheus-8                          924.7µ ± 3%    920.4µ ± 1%       ~ (p=0.699 n=6)
Query/nh_1_+_nh_1,_range_query_with_100_steps/Prometheus-8                   1.798m ± 1%    1.799m ± 1%       ~ (p=0.843 n=12)
Query/nh_1,_range_query_with_1000_steps/Mimir-8                              1.753m ± 1%    1.762m ± 3%       ~ (p=0.394 n=6)
Query/nh_1_+_nh_1,_range_query_with_1000_steps/Mimir-8                       3.915m ± 1%    3.912m ± 2%       ~ (p=0.630 n=12)
Query/nh_1,_range_query_with_1000_steps/Prometheus-8                         1.800m ± 4%    1.769m ± 2%       ~ (p=0.093 n=6)
Query/nh_1_+_nh_1,_range_query_with_1000_steps/Prometheus-8                  4.567m ± 2%    4.599m ± 1%       ~ (p=0.478 n=12)
Query/nh_100,_instant_query/Mimir-8                                          11.68m ± 2%    11.90m ± 0%       ~ (p=0.065 n=6)
Query/nh_100_+_nh_100,_instant_query/Mimir-8                                 23.85m ± 0%    23.94m ± 0%       ~ (p=0.068 n=12)
Query/nh_100,_instant_query/Prometheus-8                                     11.91m ± 1%    11.95m ± 1%       ~ (p=0.485 n=6)
Query/nh_100_+_nh_100,_instant_query/Prometheus-8                            23.53m ± 0%    23.61m ± 1%  +0.36% (p=0.033 n=12)
Query/nh_100,_range_query_with_100_steps/Mimir-8                             21.15m ± 2%    20.52m ± 3%  -2.98% (p=0.015 n=6)
Query/nh_100_+_nh_100,_range_query_with_100_steps/Mimir-8                    48.20m ± 2%    48.63m ± 3%       ~ (p=0.977 n=12)
Query/nh_100,_range_query_with_100_steps/Prometheus-8                        20.85m ± 2%    20.34m ± 3%  -2.43% (p=0.026 n=6)
Query/nh_100_+_nh_100,_range_query_with_100_steps/Prometheus-8               59.18m ± 4%    58.78m ± 4%       ~ (p=0.410 n=12)
Query/nh_100,_range_query_with_1000_steps/Mimir-8                            112.4m ± 1%    113.1m ± 1%       ~ (p=0.240 n=6)
Query/nh_100_+_nh_100,_range_query_with_1000_steps/Mimir-8                   278.4m ± 1%    281.9m ± 0%  +1.24% (p=0.033 n=12)
Query/nh_100,_range_query_with_1000_steps/Prometheus-8                       112.0m ± 1%    112.2m ± 1%       ~ (p=0.589 n=6)
Query/nh_100_+_nh_100,_range_query_with_1000_steps/Prometheus-8              453.4m ± 2%    459.7m ± 1%       ~ (p=0.160 n=12)
Query/nh_2000,_instant_query/Mimir-8                                         215.2m ± 1%    215.4m ± 0%       ~ (p=0.818 n=6)
Query/nh_2000_+_nh_2000,_instant_query/Mimir-8                               446.2m ± 0%    447.3m ± 0%       ~ (p=0.198 n=12)
Query/nh_2000,_instant_query/Prometheus-8                                    216.8m ± 1%    216.7m ± 1%       ~ (p=0.818 n=6)
Query/nh_2000_+_nh_2000,_instant_query/Prometheus-8                          439.9m ± 0%    441.0m ± 0%       ~ (p=0.219 n=12)
Query/nh_2000,_range_query_with_100_steps/Mimir-8                            422.1m ± 3%    427.5m ± 1%       ~ (p=0.180 n=6)
Query/nh_2000_+_nh_2000,_range_query_with_100_steps/Mimir-8                  999.9m ± 1%   1004.3m ± 1%  +0.44% (p=0.010 n=12)
Query/nh_2000,_range_query_with_100_steps/Prometheus-8                       416.8m ± 3%    419.7m ± 1%       ~ (p=0.132 n=6)
Query/nh_2000_+_nh_2000,_range_query_with_100_steps/Prometheus-8              1.435 ± 1%     1.438 ± 1%       ~ (p=0.590 n=12)
Query/nh_2000,_range_query_with_1000_steps/Mimir-8                            2.781 ± 2%     2.766 ± 3%       ~ (p=1.000 n=6)
Query/nh_2000_+_nh_2000,_range_query_with_1000_steps/Mimir-8                  5.850 ± 1%     5.821 ± 0%       ~ (p=0.347 n=12)
Query/nh_2000,_range_query_with_1000_steps/Prometheus-8                       2.795 ± 1%     2.771 ± 2%       ~ (p=0.937 n=6)
Query/nh_2000_+_nh_2000,_range_query_with_1000_steps/Prometheus-8             11.02 ± 1%     11.02 ± 1%       ~ (p=0.630 n=12)
Query/rate(nh_1[1m]),_instant_query/Mimir-8                                  676.1µ ± 1%    673.8µ ± 1%       ~ (p=0.818 n=6)
Query/rate(nh_1[1m]),_instant_query/Prometheus-8                             704.9µ ± 2%    703.3µ ± 1%       ~ (p=0.937 n=6)
Query/rate(nh_1[1m]),_range_query_with_100_steps/Mimir-8                     1.022m ± 1%    1.026m ± 1%       ~ (p=0.240 n=6)
Query/rate(nh_1[1m]),_range_query_with_100_steps/Prometheus-8                1.006m ± 1%    1.000m ± 1%       ~ (p=0.310 n=6)
Query/rate(nh_1[1m]),_range_query_with_1000_steps/Mimir-8                    1.929m ± 1%    1.927m ± 1%       ~ (p=0.937 n=6)
Query/rate(nh_1[1m]),_range_query_with_1000_steps/Prometheus-8               2.285m ± 2%    2.295m ± 1%       ~ (p=0.485 n=6)
Query/rate(nh_100[1m]),_instant_query/Mimir-8                                1.721m ± 1%    1.728m ± 1%       ~ (p=0.699 n=6)
Query/rate(nh_100[1m]),_instant_query/Prometheus-8                           2.012m ± 2%    2.000m ± 1%       ~ (p=0.394 n=6)
Query/rate(nh_100[1m]),_range_query_with_100_steps/Mimir-8                   31.12m ± 1%    31.27m ± 1%       ~ (p=0.240 n=6)
Query/rate(nh_100[1m]),_range_query_with_100_steps/Prometheus-8              25.38m ± 1%    25.33m ± 1%       ~ (p=0.818 n=6)
Query/rate(nh_100[1m]),_range_query_with_1000_steps/Mimir-8                  120.6m ± 1%    122.2m ± 3%  +1.33% (p=0.002 n=6)
Query/rate(nh_100[1m]),_range_query_with_1000_steps/Prometheus-8             155.3m ± 1%    154.8m ± 1%       ~ (p=0.394 n=6)
Query/rate(nh_2000[1m]),_instant_query/Mimir-8                               17.09m ± 1%    16.97m ± 1%       ~ (p=0.180 n=6)
Query/rate(nh_2000[1m]),_instant_query/Prometheus-8                          21.88m ± 2%    21.88m ± 1%       ~ (p=0.937 n=6)
Query/rate(nh_2000[1m]),_range_query_with_100_steps/Mimir-8                  607.2m ± 1%    610.4m ± 1%       ~ (p=0.485 n=6)
Query/rate(nh_2000[1m]),_range_query_with_100_steps/Prometheus-8             487.8m ± 3%    484.4m ± 1%  -0.69% (p=0.026 n=6)
Query/rate(nh_2000[1m]),_range_query_with_1000_steps/Mimir-8                  2.735 ± 1%     2.731 ± 1%       ~ (p=0.589 n=6)
Query/rate(nh_2000[1m]),_range_query_with_1000_steps/Prometheus-8             3.394 ± 0%     3.377 ± 1%  -0.51% (p=0.009 n=6)
Query/rate(nh_1[1m]),_range_query_with_10000_steps/Mimir-8                   12.98m ± 2%    13.06m ± 1%       ~ (p=0.132 n=6)
Query/rate(nh_1[1m]),_range_query_with_10000_steps/Prometheus-8              16.32m ± 1%    16.41m ± 1%       ~ (p=0.240 n=6)
Query/rate(nh_100[1m]),_range_query_with_10000_steps/Mimir-8                  1.367 ± 1%     1.372 ± 1%       ~ (p=0.065 n=6)
Query/rate(nh_100[1m]),_range_query_with_10000_steps/Prometheus-8             1.711 ± 1%     1.706 ± 1%       ~ (p=0.818 n=6)
Query/rate(nh_2000[1m]),_range_query_with_10000_steps/Mimir-8                 26.58 ± 1%     26.71 ± 1%       ~ (p=0.240 n=6)
Query/rate(nh_2000[1m]),_range_query_with_10000_steps/Prometheus-8            33.59 ± 0%     33.55 ± 0%       ~ (p=0.394 n=6)
Query/rate(nh_1[1h]),_instant_query/Mimir-8                                  1.040m ± 1%    1.050m ± 3%       ~ (p=0.065 n=6)
Query/rate(nh_1[1h]),_instant_query/Prometheus-8                             1.088m ± 2%    1.092m ± 1%       ~ (p=0.310 n=6)
Query/rate(nh_1[1h]),_range_query_with_100_steps/Mimir-8                     1.327m ± 2%    1.335m ± 1%  +0.63% (p=0.004 n=6)
Query/rate(nh_1[1h]),_range_query_with_100_steps/Prometheus-8                1.782m ± 3%    1.785m ± 1%       ~ (p=0.818 n=6)
Query/rate(nh_1[1h]),_range_query_with_1000_steps/Mimir-8                    4.459m ± 0%    4.484m ± 0%  +0.56% (p=0.009 n=6)
Query/rate(nh_1[1h]),_range_query_with_1000_steps/Prometheus-8               7.918m ± 1%    7.952m ± 1%       ~ (p=0.485 n=6)
Query/rate(nh_100[1h]),_instant_query/Mimir-8                                24.07m ± 1%    24.11m ± 2%       ~ (p=0.589 n=6)
Query/rate(nh_100[1h]),_instant_query/Prometheus-8                           23.24m ± 1%    23.08m ± 0%       ~ (p=0.065 n=6)
Query/rate(nh_100[1h]),_range_query_with_100_steps/Mimir-8                   54.05m ± 0%    54.19m ± 0%       ~ (p=0.485 n=6)
Query/rate(nh_100[1h]),_range_query_with_100_steps/Prometheus-8              87.88m ± 0%    87.83m ± 1%       ~ (p=0.937 n=6)
Query/rate(nh_100[1h]),_range_query_with_1000_steps/Mimir-8                  360.2m ± 2%    361.8m ± 1%       ~ (p=0.240 n=6)
Query/rate(nh_100[1h]),_range_query_with_1000_steps/Prometheus-8             697.3m ± 2%    696.5m ± 2%       ~ (p=0.818 n=6)
Query/rate(nh_2000[1h]),_instant_query/Mimir-8                               455.1m ± 1%    457.0m ± 1%       ~ (p=0.589 n=6)
Query/rate(nh_2000[1h]),_instant_query/Prometheus-8                          435.1m ± 2%    436.4m ± 1%       ~ (p=0.589 n=6)
Query/rate(nh_2000[1h]),_range_query_with_100_steps/Mimir-8                   1.100 ± 1%     1.098 ± 0%       ~ (p=0.240 n=6)
Query/rate(nh_2000[1h]),_range_query_with_100_steps/Prometheus-8              1.760 ± 1%     1.761 ± 1%       ~ (p=0.589 n=6)
Query/rate(nh_2000[1h]),_range_query_with_1000_steps/Mimir-8                  7.562 ± 1%     7.639 ± 2%       ~ (p=0.180 n=6)
Query/rate(nh_2000[1h]),_range_query_with_1000_steps/Prometheus-8             14.10 ± 1%     13.98 ± 1%       ~ (p=0.132 n=6)
Query/avg_over_time(nh_1[1m]),_instant_query/Mimir-8                         676.0µ ± 1%    680.7µ ± 2%       ~ (p=0.132 n=6)
Query/avg_over_time(nh_1[1m]),_instant_query/Prometheus-8                    713.5µ ± 2%    713.7µ ± 2%       ~ (p=0.937 n=6)
Query/avg_over_time(nh_1[1m]),_range_query_with_100_steps/Mimir-8            1.282m ± 2%    1.280m ± 1%       ~ (p=0.818 n=6)
Query/avg_over_time(nh_1[1m]),_range_query_with_100_steps/Prometheus-8       1.507m ± 6%    1.534m ± 1%       ~ (p=0.093 n=6)
Query/avg_over_time(nh_1[1m]),_range_query_with_1000_steps/Mimir-8           4.426m ± 1%    4.427m ± 1%       ~ (p=0.818 n=6)
Query/avg_over_time(nh_1[1m]),_range_query_with_1000_steps/Prometheus-8      7.281m ± 5%    7.298m ± 1%       ~ (p=0.699 n=6)
Query/avg_over_time(nh_100[1m]),_instant_query/Mimir-8                       2.006m ± 2%    2.001m ± 1%       ~ (p=1.000 n=6)
Query/avg_over_time(nh_100[1m]),_instant_query/Prometheus-8                  2.450m ± 1%    2.503m ± 2%  +2.14% (p=0.026 n=6)
Query/avg_over_time(nh_100[1m]),_range_query_with_100_steps/Mimir-8          56.54m ± 1%    56.47m ± 1%       ~ (p=0.937 n=6)
Query/avg_over_time(nh_100[1m]),_range_query_with_100_steps/Prometheus-8     78.64m ± 4%    77.20m ± 3%       ~ (p=0.937 n=6)
Query/avg_over_time(nh_100[1m]),_range_query_with_1000_steps/Mimir-8         372.0m ± 1%    373.4m ± 1%       ~ (p=0.699 n=6)
Query/avg_over_time(nh_100[1m]),_range_query_with_1000_steps/Prometheus-8    731.1m ± 1%    727.9m ± 2%       ~ (p=0.310 n=6)
Query/avg_over_time(nh_2000[1m]),_instant_query/Mimir-8                      22.05m ± 1%    22.01m ± 2%       ~ (p=0.937 n=6)
Query/avg_over_time(nh_2000[1m]),_instant_query/Prometheus-8                 31.26m ± 3%    31.02m ± 4%       ~ (p=0.818 n=6)
Query/avg_over_time(nh_2000[1m]),_range_query_with_100_steps/Mimir-8          1.163 ± 1%     1.164 ± 1%       ~ (p=0.485 n=6)
Query/avg_over_time(nh_2000[1m]),_range_query_with_100_steps/Prometheus-8     1.750 ± 3%     1.751 ± 2%       ~ (p=0.818 n=6)
Query/avg_over_time(nh_2000[1m]),_range_query_with_1000_steps/Mimir-8         8.157 ± 0%     8.155 ± 0%       ~ (p=1.000 n=6)
Query/avg_over_time(nh_2000[1m]),_range_query_with_1000_steps/Prometheus-8    16.53 ± 1%     16.48 ± 0%  -0.30% (p=0.015 n=6)
Query/sum_over_time(nh_1[1m]),_instant_query/Mimir-8                         669.8µ ± 1%    682.4µ ± 1%  +1.89% (p=0.002 n=6)
Query/sum_over_time(nh_1[1m]),_instant_query/Prometheus-8                    699.7µ ± 1%    708.4µ ± 1%  +1.24% (p=0.002 n=6)
Query/sum_over_time(nh_1[1m]),_range_query_with_100_steps/Mimir-8            1.062m ± 1%    1.060m ± 1%       ~ (p=0.589 n=6)
Query/sum_over_time(nh_1[1m]),_range_query_with_100_steps/Prometheus-8       1.048m ± 1%    1.046m ± 1%       ~ (p=0.589 n=6)
Query/sum_over_time(nh_1[1m]),_range_query_with_1000_steps/Mimir-8           2.430m ± 1%    2.418m ± 1%       ~ (p=0.394 n=6)
Query/sum_over_time(nh_1[1m]),_range_query_with_1000_steps/Prometheus-8      2.719m ± 1%    2.730m ± 1%       ~ (p=0.240 n=6)
Query/sum_over_time(nh_100[1m]),_instant_query/Mimir-8                       1.778m ± 1%    1.793m ± 1%  +0.81% (p=0.041 n=6)
Query/sum_over_time(nh_100[1m]),_instant_query/Prometheus-8                  2.018m ± 1%    2.036m ± 1%  +0.93% (p=0.041 n=6)
Query/sum_over_time(nh_100[1m]),_range_query_with_100_steps/Mimir-8          35.48m ± 0%    35.44m ± 1%       ~ (p=0.394 n=6)
Query/sum_over_time(nh_100[1m]),_range_query_with_100_steps/Prometheus-8     29.68m ± 2%    29.70m ± 1%       ~ (p=0.394 n=6)
Query/sum_over_time(nh_100[1m]),_range_query_with_1000_steps/Mimir-8         171.0m ± 1%    170.9m ± 1%       ~ (p=0.699 n=6)
Query/sum_over_time(nh_100[1m]),_range_query_with_1000_steps/Prometheus-8    199.7m ± 1%    199.8m ± 1%       ~ (p=0.818 n=6)
Query/sum_over_time(nh_2000[1m]),_instant_query/Mimir-8                      17.99m ± 1%    18.01m ± 1%       ~ (p=0.589 n=6)
Query/sum_over_time(nh_2000[1m]),_instant_query/Prometheus-8                 22.68m ± 1%    22.60m ± 1%       ~ (p=0.818 n=6)
Query/sum_over_time(nh_2000[1m]),_range_query_with_100_steps/Mimir-8         696.3m ± 1%    696.0m ± 1%       ~ (p=0.485 n=6)
Query/sum_over_time(nh_2000[1m]),_range_query_with_100_steps/Prometheus-8    581.1m ± 1%    582.9m ± 0%       ~ (p=0.589 n=6)
Query/sum_over_time(nh_2000[1m]),_range_query_with_1000_steps/Mimir-8         3.629 ± 0%     3.615 ± 1%       ~ (p=0.132 n=6)
Query/sum_over_time(nh_2000[1m]),_range_query_with_1000_steps/Prometheus-8    4.228 ± 1%     4.223 ± 1%       ~ (p=0.589 n=6)
Query/nh_1_/_a_1,_instant_query/Mimir-8                                      1.377m ± 1%    1.375m ± 2%       ~ (p=0.937 n=6)
Query/nh_1_/_a_1,_instant_query/Prometheus-8                                 1.403m ± 1%    1.403m ± 1%       ~ (p=0.699 n=6)
Query/nh_1_/_a_1,_range_query_with_100_steps/Mimir-8                         1.482m ± 1%    1.506m ± 2%  +1.60% (p=0.002 n=6)
Query/nh_1_/_a_1,_range_query_with_100_steps/Prometheus-8                    1.566m ± 1%    1.577m ± 1%       ~ (p=0.132 n=6)
Query/nh_1_/_a_1,_range_query_with_1000_steps/Mimir-8                        2.720m ± 5%    2.789m ± 3%       ~ (p=0.485 n=6)
Query/nh_1_/_a_1,_range_query_with_1000_steps/Prometheus-8                   3.312m ± 2%    3.364m ± 2%       ~ (p=0.132 n=6)
Query/nh_100_/_a_100,_instant_query/Mimir-8                                  14.09m ± 1%    14.19m ± 1%       ~ (p=0.065 n=6)
Query/nh_100_/_a_100,_instant_query/Prometheus-8                             14.05m ± 1%    14.03m ± 1%       ~ (p=0.589 n=6)
Query/nh_100_/_a_100,_range_query_with_100_steps/Mimir-8                     27.45m ± 1%    27.32m ± 2%       ~ (p=0.699 n=6)
Query/nh_100_/_a_100,_range_query_with_100_steps/Prometheus-8                34.80m ± 4%    35.71m ± 3%       ~ (p=0.394 n=6)
Query/nh_100_/_a_100,_range_query_with_1000_steps/Mimir-8                    154.8m ± 2%    156.5m ± 1%       ~ (p=0.240 n=6)
Query/nh_100_/_a_100,_range_query_with_1000_steps/Prometheus-8               292.4m ± 1%    292.5m ± 2%       ~ (p=0.937 n=6)
Query/nh_2000_/_a_2000,_instant_query/Mimir-8                                251.8m ± 1%    250.6m ± 1%       ~ (p=0.180 n=6)
Query/nh_2000_/_a_2000,_instant_query/Prometheus-8                           250.1m ± 0%    251.2m ± 1%       ~ (p=0.180 n=6)
Query/nh_2000_/_a_2000,_range_query_with_100_steps/Mimir-8                   568.8m ± 1%    567.4m ± 1%       ~ (p=0.937 n=6)
Query/nh_2000_/_a_2000,_range_query_with_100_steps/Prometheus-8              863.1m ± 3%    851.0m ± 2%       ~ (p=0.240 n=6)
Query/nh_2000_/_a_2000,_range_query_with_1000_steps/Mimir-8                   3.449 ± 1%     3.442 ± 1%       ~ (p=0.589 n=6)
Query/nh_2000_/_a_2000,_range_query_with_1000_steps/Prometheus-8              7.285 ± 3%     7.264 ± 1%       ~ (p=0.937 n=6)
Query/sum(nh_1),_instant_query/Mimir-8                                       818.7µ ± 1%    809.2µ ± 1%  -1.16% (p=0.041 n=6)
Query/sum(nh_1),_instant_query/Prometheus-8                                  829.6µ ± 2%    837.4µ ± 1%       ~ (p=0.065 n=6)
Query/sum(nh_1),_range_query_with_100_steps/Mimir-8                          906.9µ ± 1%    935.2µ ± 1%  +3.12% (p=0.002 n=6)
Query/sum(nh_1),_range_query_with_100_steps/Prometheus-8                     1.002m ± 2%    1.004m ± 1%       ~ (p=0.818 n=6)
Query/sum(nh_1),_range_query_with_1000_steps/Mimir-8                         1.775m ± 2%    1.861m ± 2%  +4.85% (p=0.002 n=6)
Query/sum(nh_1),_range_query_with_1000_steps/Prometheus-8                    2.259m ± 2%    2.256m ± 2%       ~ (p=0.818 n=6)
Query/sum(nh_100),_instant_query/Mimir-8                                     11.88m ± 1%    11.97m ± 1%       ~ (p=0.093 n=6)
Query/sum(nh_100),_instant_query/Prometheus-8                                11.96m ± 1%    11.92m ± 1%       ~ (p=0.589 n=6)
Query/sum(nh_100),_range_query_with_100_steps/Mimir-8                        22.07m ± 3%    21.80m ± 5%       ~ (p=0.485 n=6)
Query/sum(nh_100),_range_query_with_100_steps/Prometheus-8                   22.96m ± 4%    22.92m ± 3%       ~ (p=0.818 n=6)
Query/sum(nh_100),_range_query_with_1000_steps/Mimir-8                       118.2m ± 4%    118.1m ± 4%       ~ (p=0.937 n=6)
Query/sum(nh_100),_range_query_with_1000_steps/Prometheus-8                  161.7m ± 4%    163.6m ± 4%       ~ (p=0.818 n=6)
Query/sum(nh_2000),_instant_query/Mimir-8                                    215.1m ± 1%    215.4m ± 1%       ~ (p=0.818 n=6)
Query/sum(nh_2000),_instant_query/Prometheus-8                               216.5m ± 1%    216.6m ± 0%       ~ (p=0.818 n=6)
Query/sum(nh_2000),_range_query_with_100_steps/Mimir-8                       427.3m ± 2%    422.9m ± 2%       ~ (p=0.180 n=6)
Query/sum(nh_2000),_range_query_with_100_steps/Prometheus-8                  548.5m ± 1%    545.4m ± 2%       ~ (p=0.394 n=6)
Query/sum(nh_2000),_range_query_with_1000_steps/Mimir-8                       2.345 ± 2%     2.332 ± 1%       ~ (p=0.485 n=6)
Query/sum(nh_2000),_range_query_with_1000_steps/Prometheus-8                  4.145 ± 3%     4.164 ± 2%       ~ (p=1.000 n=6)
Query/sum_by_(l)(nh_1),_instant_query/Mimir-8                                813.3µ ± 1%    820.1µ ± 1%       ~ (p=0.180 n=6)
Query/sum_by_(l)(nh_1),_instant_query/Prometheus-8                           825.8µ ± 2%    834.9µ ± 1%       ~ (p=0.093 n=6)
Query/sum_by_(l)(nh_1),_range_query_with_100_steps/Mimir-8                   920.7µ ± 1%    941.5µ ± 1%  +2.26% (p=0.002 n=6)
Query/sum_by_(l)(nh_1),_range_query_with_100_steps/Prometheus-8              981.7µ ± 2%    986.0µ ± 1%       ~ (p=0.132 n=6)
…